Move project from local SQLite to Turso using LibSQL adapter
- Updated src/server/db.ts to use @prisma/adapter-libsql with Turso credentials
- Consolidated all Prisma Client references to use the shared instance from @/server/db
- Removed redundant src/lib/prisma.ts
- Updated next.config.ts to include serverExternalPackages for LibSQL dependencies
- Cleaned up local SQLite database files
- Updated package.json with necessary @prisma/adapter-libsql and @libsql/client dependencies
