+++ next.config.ts
@@ -17,6 +17,7 @@ const nextConfig: NextConfig = {
       },
     ],
   },
+  serverExternalPackages: ["@libsql/client", "libsql", "@prisma/adapter-libsql"],
 };
 
 export default nextConfig;
