+++ package.json
@@ -9,9 +9,9 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@libsql/client": "^0.15.15",
-    "@prisma/adapter-libsql": "^6.19.0",
-    "@prisma/client": "^6.11.1",
+    "@libsql/client": "0.17.3",
+    "@prisma/adapter-libsql": "6.19.0",
+    "@prisma/client": "6.19.0",
     "@tanstack/react-query": "^5.83.0",
     "@trpc/client": "^11.4.3",
     "@trpc/next": "^11.4.3",
@@ -22,7 +22,7 @@
     "lru-cache": "^11.1.0",
     "lucide-react": "^0.525.0",
     "next": "15.3.8",
-    "prisma": "^6.11.1",
+    "prisma": "^6.19.0",
     "react": "^19.2.0",
     "react-dom": "19.2.0",
     "zod": "^4.0.5"
