+++ src/app/(Public-route)/api/aboutme/route.ts
@@ -1,5 +1,5 @@
 import { NextRequest, NextResponse } from "next/server";
-import { prisma } from "@/lib/prisma";
+import { prisma } from "@/server/db";
 
 export async function GET() {
   try {
