+++ src/app/(Public-route)/api/blog/[slug]/route.ts
@@ -1,5 +1,5 @@
 import { NextRequest, NextResponse } from "next/server";
-import { prisma } from "@/lib/prisma";
+import { prisma } from "@/server/db";
 import { TryCatch } from "@/util/TryCatch";
 export async function PUT(request: NextRequest) {
   if (request.body === null) {
