+++ src/app/(Public-route)/api/prayer-request/route.ts
@@ -1,6 +1,6 @@
 
 import { NextRequest, NextResponse } from "next/server";
-import { prisma } from "@/lib/prisma";
+import { prisma } from "@/server/db";
 import { TryCatch } from "@/util/TryCatch";
 export async function GET() {
   const {Data, Error} = await TryCatch(prisma.prayerRequest.findMany());
