--- src/lib/prisma.ts
@@ -1,26 +1,0 @@
-import { PrismaClient } from "@prisma/client";
-import { PrismaLibSQL } from "@prisma/adapter-libsql";
-import { createClient } from "@libsql/client";
-
-const globalForPrisma = globalThis as unknown as {
-  prisma: PrismaClient | undefined;
-};
-
-const connectionString = process.env.DATABASE_URL;
-const authToken = process.env.DATABASE_AUTHTOKEN;
-
-if (!connectionString) {
-  throw new Error("DATABASE_URL is not set");
-}
-
-const libsql = createClient({
-  url: connectionString,
-  authToken: authToken,
-});
-
-const adapter = new PrismaLibSQL(libsql);
-
-export const prisma =
-  globalForPrisma.prisma ?? new PrismaClient({ adapter });
-
-if (process.env.NODE_ENV !== "production") globalForPrisma.prisma = prisma;
